requirements.txt
# FastAPI Framework
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6

# Google Vertex AI & Gemini
google-cloud-aiplatform==1.38.1
google-auth==2.25.2
google-auth-oauthlib==1.2.0
google-auth-httplib2==0.2.0

# Database
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
alembic==1.13.0

# Redis & Celery
redis==5.0.1
celery==5.3.4

# HTTP Requests
requests==2.31.0
httpx==0.25.2

# Environment & Configuration
python-dotenv==1.0.0
pydantic==2.5.2
pydantic-settings==2.1.0

# Video Processing
ffmpeg-python==0.2.0

# Authentication & Security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6

# Utilities
python-dateutil==2.8.2
pytz==2023.3

# Testing
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-cov==4.1.0
pytest-mock==3.12.0
httpx==0.25.2

# Logging & Monitoring
structlog==23.2.0
sentry-sdk==1.38.0

# Development
black==23.12.0
flake8==6.1.0
mypy==1.7.1
isort==5.13.2
